Update vip:overriddenInputs: handle relaunch
This is done client-side by clearing overriddenInputs from the
inputs map on form initialization, as it's the only place where
we already have the descriptor. Doing it server-side in
WorkflowBusiness.java:relaunch() would be possible as well, but
with more refactoring as we don't even have the application in
this context.